Starting to work with bootstrap moved old functionality from App.js to AppLoc.js AppLoc.js can still be run from a path /old  Changes to be committed: 	modified:   client/package-lock.json 	modified:   client/package.json 	new file:   client/src/css/NewStyles.module.css 	modified:   client/src/index.js 	modified:   client/src/main/App.js 	new file:   client/src/main/AppLoc.js
