+++ client/package.json
@@ -8,9 +8,11 @@
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.8.4",
+    "bootstrap": "^5.3.5",
     "cors": "^2.8.5",
     "leaflet": "^1.9.4",
     "react": "^19.1.0",
+    "react-bootstrap": "^2.10.9",
     "react-dom": "^19.1.0",
     "react-leaflet": "^5.0.0",
     "react-router-dom": "^7.5.0",
