+++ client/src/main/App.js
@@ -1,22 +1,13 @@
 import '../css/App.css';
-import { React, useState, useEffect, useRef  } from "react";
-import { MapContainer, TileLayer, Marker, Popup, useMap, LayerGroup, useMapEvents, GeoJSON, Circle } from 'react-leaflet';
+import { React, useState, useEffect, useRef } from "react";
+import { MapContainer, TileLayer, Marker, Popup, useMap  } from 'react-leaflet';
 import 'leaflet/dist/leaflet.css';
 import { Icon } from 'leaflet';
-import styles from '../css/AppStyles.module.css';
+import styles from '../css/NewStyles.module.css';
 import api from '../utils/axios'; 
 import { Outlet, Link, useNavigate } from "react-router-dom";
-import axios from "axios";
-import locationIcon from "../icons/Location-icon.svg";
-import restaurantIcon from "../icons/restaurant.svg";
-import restroomIcon from "../icons/restroom.svg";
-import shelterIcon from "../icons/shelter.svg";
-import museumIcon from "../icons/museum.svg";
-import parkIcon from "../icons/park.svg";
-import shopIcon from "../icons/shop.svg";
-import otherIcon from "../icons/other.svg";
-import e from 'cors';
-
+import 'bootstrap/dist/css/bootstrap.min.css';
+import { Navbar, Nav, Button, Dropdown, Form, FormControl, Container, Row, Col, Modal } from 'react-bootstrap';
 const LoginExpiredPrompt = ({ onClose }) => {
   return (
     <div className={styles['login-prompt-overlay']}>
@@ -34,43 +25,15 @@ const LoginExpiredPrompt = ({ onClose }) => {
   );
 };
 
-// Popup overlay
-const MapClickPopup = ({ showPopup, latlng, onSelectOption, onClose }) => {
-  const [range, setRange] = useState(100);
-  console.log("MapClickPopup", latlng, showPopup);
-  if (showPopup) {
-    return (
-      <div className={styles['popup-overlay']}>
-        <div className={styles['popup-box']}>
-          <h3>Choose an Option</h3>
-          <p>Clicked: {latlng.lng.toFixed(5)}, {latlng.lat.toFixed(5)}</p>
-          <button onClick={() => onSelectOption("neighborhood")}>Show locations in neighborhood</button>
-          <div style={{ marginTop: "10px" }}>
-            <label>Range (meters): </label>
-            <select value={range} onChange={(e) => setRange(parseInt(e.target.value))}>
-              <option value={250}>250m</option>
-              <option value={500}>500m</option>
-              <option value={750}>750m</option>
-            </select>
-            <button onClick={() => onSelectOption("range", range)} style={{ marginLeft: "10px" }}>
-              Show within range
-            </button>
-          </div>
-          <button onClick={onClose} style={{ marginTop: "10px" }}>Cancel</button>
-        </div>
-      </div>
-    );
-  }
-  return null;
-};
+
 
 function App() {
 
   const [loggedIn, setLoggedIn] = useState(sessionStorage.getItem("loginStatus") === "true");
   const [isTokenValid, setIsTokenValid] = useState(false);
   const [showLoginPrompt, setShowLoginPrompt] = useState(false);
-  //const [loginValue, setLoginValue] = useState({username: "", password: ""});
-  //const [token, setToken] = useState(sessionStorage.getItem("token") || null);
+  const [loginValue, setLoginValue] = useState({username: "", password: ""});
+  const [token, setToken] = useState(sessionStorage.getItem("token") || null);
   const navigate = useNavigate();
   const [error, setError] = useState("");
   const [location, setLocation] = useState({
@@ -81,35 +44,7 @@ function App() {
   const [loading, setLoading] = useState(false);
   const [user, setUser] = useState(null); 
   const [showAdminMenu, setShowAdminMenu] = useState(false);
-  const [locationsByCategory, setLocationsByCategory] = useState({});
-  const [visibleCategories, setVisibleCategories] = useState([]);
-  const [showLayerPanel, setShowLayerPanel] = useState(true); 
-  const [clickedLocation, setClickedLocation] = useState(null);
-  const [showOptions, setShowOptions] = useState(false);
-  const [results, setResults] = useState([]);
-  const [showHoods, setShowHoods] = useState(false); // State for checkbox
-  const [neighborhoods, setNeighborhoods] = useState([]); // State for neighborhoods
-  const [displayedLocations, setDisplayedLocations] = useState([]);
-  // Added this state to track what display mode we're in
-  const [displayMode, setDisplayMode] = useState('categories'); // 'categories', 'all', 'query'
-  const [address, setAddress] = useState("");
-  const [suggestions, setSuggestions] = useState([]);
-  const [coordinates, setCoordinates] = useState(null);
-  const [showMapClickPopup, setShowMapClickPopup] = useState(false);
-  const [geometry, setGeometry] = useState(null); // For neighborhood boundaries or range
-const [rangeCircle, setRangeCircle] = useState(null); // For the range circle
-  const BASE_URL = process.env.REACT_APP_BASE_URL || "http://localhost:5000";
-  const categoryIcons = {
-    Restrooms: new Icon({ iconUrl: restroomIcon, iconSize: [25, 25] }),
-    Restaurants: new Icon({ iconUrl: restaurantIcon, iconSize: [25, 25] }),
-    Shops: new Icon({ iconUrl: shopIcon, iconSize: [25, 25] }),
-    Shelters: new Icon({ iconUrl: shelterIcon, iconSize: [25, 25] }),
-    Parks: new Icon({ iconUrl: parkIcon, iconSize: [25, 25] }),
-    Museums: new Icon({ iconUrl: museumIcon, iconSize: [25, 25] }),
-    Others: new Icon({ iconUrl: otherIcon, iconSize: [25, 25] }),
-    // fallback/default
-    default: new Icon({ iconUrl: require("../icons/destination.png"), iconSize: [36, 36] }),
-  };
+
   const mapRef = useRef(null);
 
   
@@ -234,212 +169,52 @@ const [rangeCircle, setRangeCircle] = useState(null); // For the range circle
     return () => clearInterval(interval);
   }, []);
 
-  useEffect(() => {
-    const fetchLocations = async () => {
-      try {
-        const token = sessionStorage.getItem("token");
-        if (!token) {
-          console.warn("No token found in sessionStorage.");
-          return;
-        }
-        const res = await axios.get(`${BASE_URL}/api/locs/all`, {
-          headers: {
-            Authorization: `${token}`,
-            "Content-Type": "application/json"
-          }
-        });
-        const features = res.data.locations;
-  
-        const grouped = {};
-        features.forEach(feature => {
-          const category = feature.properties.category || "Uncategorized";
-          if (!grouped[category]) grouped[category] = [];
-          grouped[category].push(feature);
-        });
-  
-        setLocationsByCategory(grouped);
-        setVisibleCategories(Object.keys(grouped)); // Default: show all
-        //setVisibleCategories([]);
-      } catch (err) {
-        console.error("Failed to fetch locations:", err);
-      }
-    };
-  
-    fetchLocations();
-  }, []);
-
- /* const RecenterMap = ({ lat, lng })
-    const map = useMap();
-  
-    useEffect(() => {
-      map.setView([lat, lng]);
-    }, [lat, lng, map]);
-  
-    return null; // This component doesn't render anything
-  };*/
-
-  const fetchNeighborhoods = async () => {
-    try {
-      const token = sessionStorage.getItem("token");
-      const res = await axios.get(`${BASE_URL}/api/hood/neighborhoods`, {
-        headers: {
-          Authorization: `${token}`,
-          "Content-Type": "application/json",
-        },
-      });
-      setNeighborhoods(res.data.neighborhoods); // Store fetched neighborhoods
-    } catch (err) {
-      console.error("Failed to fetch neighborhoods:", err);
-    }
-  };
-
-  const MapClickHandler = ({ onMapClick }) => {
-    useMapEvents({
-      click(e) {
-        onMapClick(e.latlng); // latlng = { lat, lng }
-      },
-    });
-    return null;
-  };
-
-  const handleMapClick = (latlng) => {
-    setCoordinates(latlng)
-    setClickedLocation(latlng); // store it in state
-    setShowOptions(true);       // show a modal/popup with "in range" or "neighborhood"
-    setDisplayedLocations([]); // Clear existing locations
-  };
-
-  const geocodeAddress = async (address) => {
-    const res = await fetch(`https://nominatim.openstreetmap.org/search?q=${encodeURIComponent(address)}&format=json`);
-    const data = await res.json();
-    if (data.length > 0) {
-      return {
-        lat: parseFloat(data[0].lat),
-        lng: parseFloat(data[0].lon),
-      };
-    }
-    throw new Error("Address not found");
-  };
-
-  const handleUserChoice = async (latlng, option, range) => {
-    try {
-      setGeometry(null);
-      setRangeCircle(null);
-      setDisplayedLocations([]); // Clear existing locations
-      //console.log("User clicked:", latlng, "Option:", option, "Range:", range);
-      const token = sessionStorage.getItem("token");
-      if (option === "neighborhood") {
-        const polygon = await  axios.get(`${BASE_URL}/api/hood/position`, {
-          params: {
-            lng: latlng.lng,
-            lat: latlng.lat,
-          },
-          headers: {
-            Authorization: `${token}`,
-            "Content-Type": "application/json"
-          }
-        });
-        setGeometry(polygon.data.geometry);
-
-        const locations = await  axios.post(`${BASE_URL}/api/locs/area`, { coordinates: polygon.data.geometry.coordinates }, {
-          headers: {
-            Authorization: `${token}`,
-            "Content-Type": "application/json"
-          }
-        });
-        //console.log("Locations in neighborhood:", locations.data);
-        const features = locations.data.locations;
-  
-        const grouped = {};
-        features.forEach(feature => {
-          const category = feature.properties.category || "Uncategorized";
-          if (!grouped[category]) grouped[category] = [];
-          grouped[category].push(feature);
-        });
-        setDisplayedLocations(grouped);
-        setVisibleCategories(Object.keys(grouped));
-        setDisplayMode('query');
-      } else if (option === "range") {
-        setRangeCircle({
-          center: [latlng.lat, latlng.lng],
-          radius: range,
-        });
-        const locations = await axios.post(`${BASE_URL}/api/locs/range`, {coordinates: [latlng.lng, latlng.lat], range: range}, {
-          headers: {
-            Authorization: `${token}`,
-            "Content-Type": "application/json"
-          }
-        });
-        //console.log("Locations in range:", locations.data);          
-        const features = locations.data.locations;
-  
-        const grouped = {};
-        features.forEach(feature => {
-          const category = feature.properties.category || "Uncategorized";
-          if (!grouped[category]) grouped[category] = [];
-          grouped[category].push(feature);
-        });
-        setDisplayedLocations(grouped);
-        setVisibleCategories(Object.keys(grouped));
-        setDisplayMode('query');
-      }
-    } catch (error) {
-      console.error("Error fetching locations:", error);
-    }
-  };
-
-  // Fetch suggestions from Nominatim
-  const fetchSuggestions = async (query) => {
-    if (!query) {
-      setSuggestions([]);
-      return;
-    }
-
-    try {
-      const response = await fetch(
-        `https://nominatim.openstreetmap.org/search?q=${encodeURIComponent(
-          query
-        )}&format=json&addressdetails=1&limit=5`
+  const getHeader = () => {
+    if (loggedIn) {
+      return (
+        <Navbar bg="secondary" variant="dark" expand="lg" className="w-100">
+          <Navbar.Brand className="text-white">Welcome</Navbar.Brand>
+          <Nav className="ml-auto align-items-center">
+            <Dropdown>
+              <Dropdown.Toggle variant="link" className="text-white text-decoration-none mr-3">
+                Responders
+              </Dropdown.Toggle>
+              <Dropdown.Menu>
+                <Dropdown.Item href="#action1">Responder 1</Dropdown.Item>
+                <Dropdown.Item href="#action2">Responder 2</Dropdown.Item>
+                <Dropdown.Item href="#action3">Responder 3</Dropdown.Item>
+              </Dropdown.Menu>
+            </Dropdown>
+            <Button
+              variant="link"
+              onClick={handleLogout}
+              className="text-white text-decoration-none mr-3"
+            >
+              Logout
+            </Button>
+            {user?.role === "Admin" && (
+              <Link to="/admin" className="text-white text-decoration-none">
+                Admin
+              </Link>
+            )}
+          </Nav>
+        </Navbar>
       );
-      const data = await response.json();
-      setSuggestions(data);
-    } catch (error) {
-      console.error("Error fetching suggestions:", error);
+    } else {
+      return (
+        <Navbar bg="secondary" variant="dark" expand="lg" className="w-100">
+          <Navbar.Brand className="text-white">Welcome</Navbar.Brand>
+          <Nav className="ml-auto align-items-center">
+            <Link to="/login" className="text-white text-decoration-none mr-3">
+              Login
+            </Link>
+            <Link to="/register" className="text-white text-decoration-none ml-3">
+              Register
+            </Link>
+          </Nav>
+        </Navbar>
+      );
     }
-  };
-
-  // Handle address input change
-  const handleCompleteAddress = async (address) => {
-    //const address = e.target.value;
-    if (address) {
-      try {
-        const coords = await geocodeAddress(address);
-        setClickedLocation(coords);
-        setShowOptions(true); // trigger popup for range or neighborhood
-        setShowMapClickPopup(true); // Show the map click popup
-      } catch (err) {
-        alert("Failed to find address.");
-      }
-    }
-    setAddress(address);
-    fetchSuggestions(address);
-  };
-
-   // Handle address input change
-   const handleInputChange = (e) => {
-    const value = e.target.value;
-    setAddress(value);
-    fetchSuggestions(value);
-  };
-
-  // Handle address selection
-  const handleSelect = (suggestion) => {
-    setAddress(suggestion.display_name);
-    setCoordinates({
-      lat: parseFloat(suggestion.lat),
-      lng: parseFloat(suggestion.lon),
-    });
-    setSuggestions([]); // Clear suggestions after selection
   };
 
 const handleTokenExpiration = () => {
@@ -480,20 +255,6 @@ useEffect(() => {
   return () => api.interceptors.response.eject(interceptor);
 }, []);
 
-useEffect(() => {
-  const saved = localStorage.getItem("visibleCategories");
-  if (saved) {
-    setVisibleCategories(JSON.parse(saved));
-  }
-}, []);
-
-const handleCategoryToggle = (category, checked) => {
-  const updated = checked
-    ? [...visibleCategories, category]
-    : visibleCategories.filter((c) => c !== category);
-  setVisibleCategories(updated);
-  localStorage.setItem("visibleCategories", JSON.stringify(updated));
-};
 
 const handleLogout = () => {
   console.log("Logging out...");
@@ -509,53 +270,10 @@ const handleLogout = () => {
   window.location.reload();
 };
 
-// Handle clearing locations from map
-const handleClearLocations = () => {
-  setDisplayedLocations([]);
-  setVisibleCategories([]);
-  setDisplayMode('none');
-};
-
-// Handle showing all locations
-const handleShowAllLocations = async () => {
-  try {
-    const token = sessionStorage.getItem("token");
-    const res = await axios.get(`${BASE_URL}/api/locs/all`, {
-      headers: {
-        Authorization: `${token}`,
-        "Content-Type": "application/json",
-      },
-    });
-    const features = res.data.locations;
-  
-        const grouped = {};
-        features.forEach(feature => {
-          const category = feature.properties.category || "Uncategorized";
-          if (!grouped[category]) grouped[category] = [];
-          grouped[category].push(feature);
-        });
-        setDisplayedLocations(grouped);
-        setVisibleCategories(Object.keys(grouped));
-        setDisplayMode('all');
-    // Clear category selection when displaying all locations
-    //setVisibleCategories([]);
-  } catch (err) {
-    console.error("Failed to fetch all locations:", err);
-  }
-};
-
-const getFilteredLocations = (locations) => {
-  if (!locations) return {};
-  
-  const filtered = {};
-  Object.keys(locations).forEach(category => {
-    if (visibleCategories.includes(category)) {
-      filtered[category] = locations[category];
-    }
-  });
-  
-  return filtered;
-};
+  const customIcon = new Icon({
+    iconUrl: require("../icons/destination.png"),
+    iconSize: [36,36]
+  })
 
   const fetchAddress = async (lat, lon) => {
     const response = await fetch(
@@ -572,343 +290,77 @@ const getFilteredLocations = (locations) => {
   };
 
   return (
-    <div className={styles.container}>
+    <Container fluid className="p-0">
       {showLoginPrompt && <LoginExpiredPrompt onClose={handleCloseLoginPrompt} />}
-      <div className={styles.content}>
-        <div className={styles.header}>
-        <nav className="App-navigation">
-          <label>
-            <input
-              type="checkbox"
-              checked={showHoods}
-              onChange={(e) => {
-                setShowHoods(e.target.checked);
-                if (e.target.checked) fetchNeighborhoods(); // Fetch neighborhoods when checked
-              }}
-            />
-            Show Qtrs <a>  </a>
-          </label>
-          {!loggedIn && ( // Show these links only when not logged in
-            <>
-              <a className="App-link" href="/register">Register</a>
-              <a>/</a>
-              <a className="App-link" href="/login">Login</a>
-              <a>/</a>
-            </>
-          )}
-          {loggedIn && ( // Show these links only when logged in
-            <>
-            <button 
-              className="App-link" 
-              onClick={handleLogout}
-              style={{ background: 'none', border: 'none', cursor: 'pointer', font: 'inherit', color: 'inherit' }}
-            >
-              Logout
-            </button>
-          </>
-          )}
-          {user?.role === "Admin" && (
-            <div className={styles.adminMenuContainer}>
-            <button 
-              onClick={() => setShowAdminMenu(prev => !prev)} 
-              className={styles.adminenuToggle}
-            >
-            Admin Activities ▼
-          </button>
-          {showAdminMenu && (
-            <div className={styles.adminMenuDropdown}>
-              <Link to="/">Home</Link>
-              <Link to="/locations">Manage Locations</Link>
-              <Link to="/admin">Manage Users</Link>
-              <Link to="/hoods">Manage Neighborhoods</Link>
-            </div>
-          )}
-          </div>
-        )}
-          <span style={{ color: '#61dafb', marginLeft: '5px' }}>&copy;</span>
-
-        </nav>
-        {loggedIn && (
-          <>
-          {/* Address Input */}
-          <input
-            type="text"
-            value={address}
-            onChange={handleInputChange} // Update the state on change
-            onKeyDown={(e) => {
-              if (e.key === "Enter") {
-                handleCompleteAddress(address); // Handle the complete address on Enter
-              }
-            }}
-            placeholder="Enter an address"
-            className={styles.addressInput}
-          />
-          {/* Suggestions Dropdown */}
-          {suggestions.length > 0 && (
-            <div className={styles.suggestionsDropdown}>
-              {suggestions.map((suggestion) => (
-                <div
-                  key={suggestion.place_id}
-                  className={styles.suggestion}
-                  onClick={() => handleSelect(suggestion)}
-                >
-                  {suggestion.display_name}
-                </div>
-              ))}
-            </div>
-          )}
-          <button 
-            className={styles.mapClickButton}
-            onClick={() => setShowMapClickPopup(true)} >  
-          Click Here & On Map
-          </button>
-          </>
-        )}
-        </div>
-        <div className={styles.middle}>
-        {/*<div className={styles.layerPanelWrapper}>
-          
-        </div>*/}
-        <div className={styles.mapControls}>
-        <button 
-            onClick={() => setShowLayerPanel(prev => !prev)} 
-            className={styles.layerToggleButton}
-          >
-            {showLayerPanel ? 'Hide Layers' : 'Show Layers'}
-          </button>
-
-          {showLayerPanel && (
-            <div className={styles.legend}>
-            {Object.keys(locationsByCategory).map((category) => (
-              <label key={category} className={styles.legendItem}>
-                <input
-                  type="checkbox"
-                  checked={visibleCategories.includes(category)}
-                  onChange={(e) => handleCategoryToggle(category, e.target.checked)}
-                />
-                <img src={categoryIcons[category]?.options.iconUrl} alt="" width={20} style={{ marginRight: '5px' }} />
-                {category}
-              </label>
-            ))}
-          </div>
-          )}
-        <button
-          onClick={handleClearLocations}
-          className={styles.layerToggleButton}
-        >
-          Clear Locations
-        </button>
-        <button onClick={handleShowAllLocations} className={styles.layerToggleButton}>
-            Show All Locations
-          </button>          
-        </div>
-        <button
-          className={styles.recenterButton}
-          onClick={async() => {
-            const address = await fetchAddress(location.latitude, location.longitude);
-            if (address) { alert(`Current location: ${address}`); 
-            } else {
-              alert("Address not found.");
-            }
-            // Recenter the map to the user's current location  
-            const map = mapRef.current;
-            if (map && location.latitude && location.longitude) {
-              map.setView([location.latitude, location.longitude], 13); // Center the map on the user's current location
-            } else {
-              alert("Address Can not be set on map.");
-            }
+      <Row className="m-0" style={{ height: "100vh" }}>
+        {/* Left Box */}
+        <Col
+          md={3} // 30% of the screen width
+          className="p-3"
+          style={{
+            backgroundColor: "#6c757d", // Same color as the header
+            color: "white", // White text
+            height: "100vh", // Full height of the viewport
           }}
-        >          
-        </button>
-        <MapContainer 
-          center={[location.latitude, location.longitude]}
-          zoom={13}
-          ref={mapRef} // 👈 expose map for recentering          
         >
-          <TileLayer
-            attribution='&copy; <a href="https://osm.org/copyright">OpenStreetMap</a>  | <a href="https://www.flaticon.com/free-icons/destination" title="destination icons">Destination icons created by Flat Icons - Flaticon</a>'
-            url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-          />  
+          <h5>Active Incidents</h5>
+          <p>תהיה טבלה של אירועים</p>
+          {/* Add more content here */}
+        </Col>
+  
+        {/* Main Content */}
+        <Col md={9} className="p-0">
+          {/* Header */}
+          <Row className="m-0">
+            <Col className="p-0">
+              {getHeader()}
+            </Col>
+          </Row>
+  
+          {/* Map Section */}
+          <Row className="m-0" style={{ height: "calc(100vh - 56px)" }}>
+            {/* Subtract the height of the header (56px is the default height of a Bootstrap Navbar) */}
+            <Col className="p-0">
+              <MapContainer
+                center={[location.latitude, location.longitude]}
+                zoom={13}
+                style={{ height: "100%" }}
+                ref={mapRef}
+              >
+                <TileLayer
+                  attribution='&copy; <a href="https://osm.org/copyright">OpenStreetMap</a> | <a href="https://www.flaticon.com/free-icons/destination" title="destination icons">Destination icons created by Flat Icons - Flaticon</a>'
+                  url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+                />
+                {/* Add markers or other map components here */}
 
-          {/* Render neighborhood boundaries */}
-          {geometry && (
-            <GeoJSON
-              data={geometry}
-              style={{
-                color: "blue", // Border color
-                weight: 2, // Border thickness
-                fillColor: "blue", // Fill color
-                fillOpacity: 0.2, // Translucent fill
-              }}
-            />
-          )}
-
-          {/* Render range circle */}
-          {rangeCircle && (
-            <Circle
-              center={rangeCircle.center}
-              radius={rangeCircle.radius}
-              pathOptions={{
-                color: "red", // Circle border color
-                fillColor: "red", // Circle fill color
-                fillOpacity: 0.2, // Translucent fill
-              }}
-    />
-  )}   
-            {/*<RecenterMap lat={location.latitude} lng={location.longitude} />*/}
-            <MapClickHandler onMapClick={handleMapClick} /> 
-
-            {/* Render neighborhoods if "Show Hoods" is checked */}
-            {showHoods &&
-              neighborhoods.map((hood, idx) => (
-                <GeoJSON
-                  key={idx}
-                  data={hood}
-                  style={{
-                    color: "blue", // Border color
-                    weight: 2, // Border thickness
-                    fillColor: "blue", // Fill color
-                    fillOpacity: 0.2, // Translucent fill
+                {/* Sniper Scope Button */}
+                <button
+                  className={styles.recenterButton}
+                  onClick={async () => {
+                    const address = await fetchAddress(location.latitude, location.longitude);
+                    if (address) {
+                      alert(`Current location: ${address}`);
+                    } else {
+                      alert("Address not found.");
+                    }
+                    // Recenter the map to the user's current location
+                    const map = mapRef.current;
+                    if (map && location.latitude && location.longitude) {
+                      map.setView([location.latitude, location.longitude], 13); // Center the map on the user's current location
+                    } else {
+                      alert("Address cannot be set on the map.");
+                    }
                   }}
-                />
-              ))}              
-           {/* Render markers based on display mode */}
-          {/* 1. Show category markers when in categories mode */}
-          {displayMode === 'categories' && Object.entries(locationsByCategory).map(([category, features]) => (
-            visibleCategories.includes(category) && (
-              <LayerGroup key={category}>
-                {features.map((feature, idx) => (
-                  <Marker
-                    key={`${category}-${idx}`}
-                    position={[feature.geometry.coordinates[1], feature.geometry.coordinates[0]]}
-                    icon={categoryIcons[category] || categoryIcons.default}
-                  >
-                    <Popup>
-                      <div>
-                        <p><strong>{feature.properties.loc_name}</strong></p>
-                        {feature.properties.photo && (
-                          <img className={styles.popupImg}
-                            src={`${process.env.PUBLIC_URL}/images/${feature.properties.photo}`}
-                            alt={feature.properties.loc_name}
-                            onError={(e) => {
-                              console.error("Image failed to load:", e);
-                              e.target.onerror = null;
-                            }}
-                          />
-                        )}
-                        <p><strong>Description:</strong> {feature.properties.description}</p>
-                        {feature.properties.address && (
-                          <p>
-                            <strong>Address:</strong> {feature.properties.address}
-                          </p>
-                        )}
-                        {feature.properties.phone && (
-                          <p>
-                            <strong>Phone:</strong> {feature.properties.phone}
-                          </p>
-                        )}
-                        {feature.properties.email && (
-                          <p>
-                            <strong>Email:</strong> {feature.properties.email}
-                          </p>
-                        )}
-                        {feature.properties.Site && (
-                          <p>
-                            <strong>Site:</strong> {feature.properties.Site}
-                          </p>
-                        )}
-                        {feature.properties.category && (
-                          <p>
-                            <strong>Category:</strong> {feature.properties.category}
-                          </p>
-                        )}
-                      </div>
-                    </Popup>
-                  </Marker>
-                ))}
-              </LayerGroup>
-            )
-          ))}
-            
-          {/* 2. Show all locations or query results when in those modes */}
-          {(displayMode === 'all' || displayMode === 'query') && Object.entries(getFilteredLocations(displayedLocations)).map(([category, features]) => (
-            <LayerGroup key={category}>
-              {features.map((feature, idx) => (
-            <Marker
-            key={`${category}-${idx}`}
-              position={[
-                feature.geometry.coordinates[1],
-                feature.geometry.coordinates[0]
-              ]}
-              icon={categoryIcons[feature.properties?.category] || categoryIcons.default}
-            >
-            <Popup>
-              <div>
-                <p><strong>{feature.properties.loc_name}</strong></p>
-                {feature.properties.photo && (
-                  <img className={styles.popupImg}
-                    src={`${process.env.PUBLIC_URL}/images/${feature.properties.photo}`}
-                    alt={feature.properties.loc_name}
-                    onError={(e) => {
-                      console.error("Image failed to load:", e);
-                      e.target.onerror = null;
-                    }}
-                  />
-                )}
-                <p><strong>Description:</strong> {feature.properties.description}</p>
-                {feature.properties.address && (
-                  <p>
-                    <strong>Address:</strong> {feature.properties.address}
-                  </p>
-                )}
-                {feature.properties.phone && (
-                  <p>
-                    <strong>Phone:</strong> {feature.properties.phone}
-                  </p>
-                )}
-                {feature.properties.email && (
-                  <p>
-                    <strong>Email:</strong> {feature.properties.email}
-                  </p>
-                )}
-                {feature.properties.Site && (
-                  <p>
-                    <strong>Site:</strong> {feature.properties.Site}
-                  </p>
-                )}
-                {feature.properties.category && (
-                  <p>
-                    <strong>Category:</strong> {feature.properties.category}
-                  </p>
-                )}
-              </div>
-            </Popup>
-            </Marker>
-          ))}
-          </LayerGroup>
-        ))}        
-      </MapContainer>
-      {showOptions && clickedLocation && (
-        <MapClickPopup 
-          showPopup={showMapClickPopup}
-          latlng={clickedLocation}
-          onSelectOption={(option, range) => {
-            handleUserChoice(clickedLocation, option, range);
-            setShowOptions(false);
-            setShowMapClickPopup(false);
-          }}
-          onClose={() => {
-            setShowOptions(false);
-            setShowMapClickPopup(false);
-          }}
-        />
-      )}
-        </div>
-        <div className={styles.footer}></div>
-      </div>      
-    </div>
-    
-    
+                >
+                  {/* Sniper Scope Icon */}
+                  <div className={styles.sniperScope}></div>
+                </button>
+              </MapContainer>
+            </Col>
+          </Row>
+        </Col>
+      </Row>
+    </Container>
   );
 }
